drop module towhee.dc and add  towhee.dc as a factory method
