+++ tests/unittests/functional/test_entity.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import unittest
-from towhee.functional.entity import Entity
+from towhee import Entity
 
 
 class TestEntity(unittest.TestCase):
